% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{add_arg}
\alias{add_arg}
\title{Add a command line argument to configuration}
\usage{
add_arg(
  name,
  type = "character",
  doc = "",
  nargs = 1,
  req = FALSE,
  short = NULL,
  options = NULL,
  default = NULL
)
}
\arguments{
\item{name}{argument name to be used on command line and in final
use '.' in n ame to indicate the nest levels. e.g.,
'arg_group1.color' will be structured as arg_group1$color.
there is no limit to how many levels can be specified.}

\item{type}{data type or class of argument, or 'file' or 'path'. If type
is 'file' or 'path', existence of such file or path will be
tested. hence, not to be used for files to be created (output)}

\item{doc}{help string}

\item{nargs}{number of expected arguments. use '+' for unknown number.
default = 1}

\item{req}{logical; when TRUE, argument will be required and if not
provided, script will fail. default = FALSE}

\item{short}{short name. short character string to use as an alternative
argument identifier on command line. default = NULL}

\item{options}{vector of valid options for this argument. script will fail
if value provided does not match any of these options.
default = NULL (argument value may be any value satisfying}

\item{default}{default argument value; default = NULL}
}
\description{
Create an argument object with desired settings and add to CLI configuration
}
