% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{add_dependent_req}
\alias{add_dependent_req}
\title{Add a dependent or conditional requirement to an argument object in CLI}
\usage{
add_dependent_req(arg_not_null, req_arg)
}
\arguments{
\item{arg_not_null}{name of argument to check for null to determine
whether \code{req_arg} is required}

\item{req_arg}{name of argument to set as required when
\code{arg_not_null} is not null}
}
\description{
Add a requirement to CLI object that an argument is only required when
a certain other argument is NOT NULL.
}
