% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{add_required_choice}
\alias{add_required_choice}
\title{Add a requirement to CLI object for user to choose between two or more
arguments or sets of arguments}
\usage{
add_required_choice(choice_id, arg_choices)
}
\arguments{
\item{choice_id}{character; ID for the required choice. e.g., 'INPUT'}

\item{arg_choices}{list with two items, where each item is a vector of one
or more arguments and the user must provide either ALL
arguments in list item 1 or ALL items in list argument 2}
}
\description{
After adding all individual arguments to CLI object, set a requirement for
user to use one option or set of options OR the other.
}
