% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{args_to_nested_list}
\alias{args_to_nested_list}
\title{Create nested list from flat command line argument list}
\usage{
args_to_nested_list(args)
}
\arguments{
\item{args}{single-level, named list of command line arguments}
}
\value{
nested list
}
\description{
Parse argument list names to determine how list should be nested. Each '.'
character in an argument name indicates that an inner list should be
created. e.g., list item name 'arg_group1.color.hot' will result in an
inner list structured \code{arg_group1$color$hot}
}
