% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{build_arg_help}
\alias{build_arg_help}
\title{Format argument doc string for help}
\usage{
build_arg_help(arg, longest_arg = NULL)
}
\arguments{
\item{arg}{argument object}

\item{longest_arg}{integer; length of the longest argument name, used for
justification; default = NULL}
}
\value{
formatted argument doc string
}
\description{
Format argument doc string for help
}
