% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{check_dependent_requirements}
\alias{check_dependent_requirements}
\title{Check that dependent requirements are satisfied}
\usage{
check_dependent_requirements(cli, args)
}
\arguments{
\item{cli}{CLI object}

\item{args}{named, flat list of command line arguments}
}
\description{
If an argument is required when another one provided and said argument is
not also provided, print error message and exit with non-zero exit status
}
