% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{check_required_args}
\alias{check_required_args}
\title{Check for existence of required arguments}
\usage{
check_required_args(cli, args)
}
\arguments{
\item{cli}{CLI object}

\item{args}{flat (not nested) argument list provided on command line}
}
\description{
If any required arguments are missing from parsed argument list,
print error message and quit with non-zero exit status
}
