% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{check_required_choices}
\alias{check_required_choices}
\title{Check that required choices are made properly}
\usage{
check_required_choices(cli, args)
}
\arguments{
\item{cli}{CLI object}

\item{args}{named, flat list of command line arguments}
}
\description{
For each item in \code{cli$req_choices}, make sure that one of the two
sets of choices is provided. If not, print error message and quit with non-
zero exit status
}
