\name{CLI}
\alias{CLI}
\title{Package Specific Command Line Interface}
\usage{
  CLI(default = NULL, ARGS = commandArgs(TRUE), ...,
    package = NULL)
}
\arguments{
  \item{default}{command}

  \item{ARGS}{list of parsed arguments passed to the CLI
  entry point.}

  \item{...}{extra arguments passed to the package's CLI
  function.}

  \item{package}{name of the package that define the CLI
  entry points}
}
\description{
  Package Specific Command Line Interface
}

