% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{configure_cli}
\alias{configure_cli}
\title{Initialize CLI configuration object}
\usage{
configure_cli()
}
\description{
Create CLI object to hold required and optional arguments, required choices,
and depending requirements
}
