% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{fill_in_defaults}
\alias{fill_in_defaults}
\title{Fill in argument list with defaults for arguments not provided on
command line}
\usage{
fill_in_defaults(arglist, cli)
}
\arguments{
\item{arglist}{named, flat list of command line arguments}

\item{cli}{CLI object}
}
\description{
Fill in argument list with defaults for arguments not provided on
command line
}
