% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{get_longest_str}
\alias{get_longest_str}
\title{Get longest string in vector of strings}
\usage{
get_longest_str(strings)
}
\arguments{
\item{strings}{vector of character strings}
}
\value{
length of longest character string in vector
}
\description{
Get longest string in vector of strings
}
