% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{map_arg_names}
\alias{map_arg_names}
\title{Create a map of full argument names to their short alternatives}
\usage{
map_arg_names(cli)
}
\arguments{
\item{cli}{CLI object}
}
\value{
list of full argument names, named by their short names
}
\description{
Create a map of full argument names to their short alternatives
}
