% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{merge_lists}
\alias{merge_lists}
\title{Merge two lists}
\usage{
merge_lists(list1, list2)
}
\arguments{
\item{list1}{the first list to be merged}

\item{list2}{the second list to be merged}
}
\description{
Recursively combine two lists, returning the union of all list levels
}
