% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{nest_arg}
\alias{nest_arg}
\title{Internal function to parse argument name and create nested list using '.'
character as an indication of list levels}
\usage{
nest_arg(arg)
}
\arguments{
\item{arg}{list containing argument value and named by arg name}
}
\description{
Internal function to parse argument name and create nested list using '.'
character as an indication of list levels
}
