% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parse_cl}
\alias{parse_cl}
\title{Parse and validate command-line input}
\usage{
parse_cl(cli, log = FALSE, nest = TRUE)
}
\arguments{
\item{cli}{command-line interface configuration, built with any
combination of argument objects}

\item{log}{logical; when TRUE, final arguments will be logged in the
console}

\item{nest}{logical; when TRUE, arguments with names including a '.' will
be nested accordingly; e.g., 'arg_group1.color = red' will be
in final argument list as arg_group1$color = red}
}
\value{
named list of all valid arguments
}
\description{
Create an argument list and validate according to specifications in each CLI
argument object.
}
