% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{render_service}
\alias{render_service}
\title{Renders Service Report File}
\usage{
render_service(name, ..., params = NULL, output_dir = ".", quiet = NULL,
  envir = parent.frame(), package = topenv(parent.frame()),
  .extra.files = NULL)
}
\arguments{
\item{name}{name of the service.
It can be the name of the script that implements the service or the name of the directory
that contains it.}

\item{...}{report parameters checked against their specifications defined in the report YAML header file
and passed down to \code{\link[rmarkdown]{render}}.}

\item{params}{report parameters passed as a named list, typically from command line arguments.
If \var{params} is not \code{NULL}, then any argument passed in \var{...} is ignored and will most likely
result in an error from \link[rmarkdown:render]{rmarkdown::render}.}

\item{output_dir}{Output directory, where to generate the report.
The default is to generate the report in the current working directory.}

\item{quiet}{logical that indicates if the report should be generated quietly.
If \code{NULL} then only an overview of the script parameters is shown.}

\item{envir}{environment where the report is rendered. See \code{\link[rmarkdown]{render}}.}

\item{package}{name of the package where to look for the givne service.}

\item{.extra.files}{vector of paths to extra files to copy to the output directory.}
}
\description{
Generates a service report.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
