% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{service_files}
\alias{service_files}
\title{Path to Report File}
\usage{
service_files(name, all = FALSE, full.names = FALSE,
  package = topenv(parent.frame()))
}
\arguments{
\item{name}{name of the service.
It can be the name of the script that implements the service or the name of the directory
that contains it.}

\item{all}{logical that indicates if all the files associated with the service should be
returned.}

\item{full.names}{logical that indicates if the full path to the services file(s) should
be returned or only the path relative to the service base directory.
The base directory itself is always returned as an attribute ('path').}

\item{package}{name of the package where to look for the givne service.}
}
\description{
Path to Report File
}
