% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{structure_arg_list}
\alias{structure_arg_list}
\title{Parse command line arguments and build one-level argument list}
\usage{
structure_arg_list(args, cli)
}
\arguments{
\item{args}{list of arguments read directly from command line}

\item{cli}{CLI object}
}
\value{
structured argument linst
}
\description{
TODO: LOOK AT THIS FUNCTION MORE CLOSELY TO GET BETTER DESCRIPTION
}
