% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{usage_detail}
\alias{usage_detail}
\title{Build full usage including all argument documentation}
\usage{
usage_detail(cli)
}
\arguments{
\item{cli}{CLI object}
}
\value{
full usage string
}
\description{
Compile usage summary AND all argument documentation
}
