% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{usage_summary}
\alias{usage_summary}
\title{Build usage summary}
\usage{
usage_summary(cli)
}
\arguments{
\item{cli}{CLI object}
}
\value{
character string
}
\description{
Construct single-line script usage summary
}
