% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{validate_arg}
\alias{validate_arg}
\title{Validate provided argument using its corresponding CLI argument object}
\usage{
validate_arg(cli_arg, arg_val)
}
\arguments{
\item{cli_arg}{argument object from CLI object}

\item{arg_val}{value provided for argument}
}
\description{
Use specifications in CLI argument object to validate provided argument
}
