% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{validate_args}
\alias{validate_args}
\title{Validate all arguments}
\usage{
validate_args(cli, args)
}
\arguments{
\item{cli}{CLI object}

\item{args}{flat, named list of arguments provided on command line}
}
\description{
Check all arguments provided against the specifications in the corresponding
CLI object argument
}
