% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLI.R
\name{yaml.docopt}
\alias{yaml.docopt}
\title{Extract Docopt String from R Markdown YAML Header}
\usage{
yaml.docopt(file)
}
\arguments{
\item{file}{rmarkdown file}
}
\value{
string
}
\description{
Extract Docopt String from R Markdown YAML Header
}
