% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{render_service}
\alias{render_service}
\title{Renders Service Report File}
\usage{
render_service(service, ..., args = NULL, work_dir = ".", save = FALSE,
  envir = parent.frame(), package = NULL)
}
\arguments{
\item{...}{report parameters checked against their specifications defined in the report YAML header file
and passed down to \code{\link[rmarkdown]{render}}.}

\item{args}{report parameters passed as a character vector, typically from command line arguments.
If \var{args} is not \code{NULL}, then any argument passed in \var{...} is ignored.}

\item{work_dir}{path to working directory}

\item{envir}{environment where the report is rendered. See \code{\link[rmarkdown]{render}}.}

\item{help}{logical that requests the service man page to be shown.
If \code{TRUE}, the service is not run at all, and only show the man page before exiting.}
}
\description{
Generates a service report.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
