% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{render_service}
\alias{render_service}
\title{Renders Service Report File}
\usage{
render_service(name, ..., args = NULL, work_dir = ".", save = FALSE,
  envir = parent.frame(), package = topns_name())
}
\arguments{
\item{name}{name of the service.
It can be the name of the script that implements the service or the name of the directory
that contains it.}

\item{...}{report parameters checked against their specifications defined in the report YAML header file
and passed down to \code{\link[rmarkdown]{render}}.}

\item{args}{report parameters passed as a character vector, typically from command line arguments.
If \var{args} is not \code{NULL}, then any argument passed in \var{...} is ignored.}

\item{work_dir}{path to working directory}

\item{save}{logical or character string that specify if or where the result list should be saved.}

\item{envir}{environment where the report is rendered. See \code{\link[rmarkdown]{render}}.}

\item{package}{name of the package where to look for the givne service.}
}
\description{
Generates a service report.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
