% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{render_service}
\alias{render_service}
\title{Renders Service Report File}
\usage{
render_service(name, ..., args = NULL, outdir = ".",
  envir = parent.frame(), package = topenv(parent.frame()),
  .extra.files = NULL)
}
\arguments{
\item{name}{name of the service.
It can be the name of the script that implements the service or the name of the directory
that contains it.}

\item{...}{report parameters checked against their specifications defined in the report YAML header file
and passed down to \code{\link[rmarkdown]{render}}.}

\item{args}{report parameters passed as a character vector, typically from command line arguments.
If \var{args} is not \code{NULL}, then any argument passed in \var{...} is ignored.}

\item{outdir}{path to working directory}

\item{envir}{environment where the report is rendered. See \code{\link[rmarkdown]{render}}.}

\item{package}{name of the package where to look for the givne service.}

\item{.extra.files}{vector of paths to extra files to copy to the output directory.}
}
\description{
Generates a service report.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
