% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{yaml.docopt}
\alias{yaml.docopt}
\title{Extract Docopt String from R Markdown YAML Header}
\usage{
yaml.docopt(file)
}
\arguments{
  \item{file}{rmarkdown file}
}
\value{
  string
}
\description{
  Extract Docopt String from R Markdown YAML Header
}

